stage_control_tab -> StageControlTab
