--- src/aslm/view/main_window_content/camera_display/camera_view/camera_notebook.py
@@ -1,72 +1,0 @@
-# Copyright (c) 2021-2022  The University of Texas Southwestern Medical Center.
-# All rights reserved.
-
-# Redistribution and use in source and binary forms, with or without
-# modification, are permitted for academic and research use only (subject to the limitations in the disclaimer below)
-# provided that the following conditions are met:
-
-#      * Redistributions of source code must retain the above copyright notice,
-#      this list of conditions and the following disclaimer.
-
-#      * Redistributions in binary form must reproduce the above copyright
-#      notice, this list of conditions and the following disclaimer in the
-#      documentation and/or other materials provided with the distribution.
-
-#      * Neither the name of the copyright holders nor the names of its
-#      contributors may be used to endorse or promote products derived from this
-#      software without specific prior written permission.
-
-# NO EXPRESS OR IMPLIED LICENSES TO ANY PARTY'S PATENT RIGHTS ARE GRANTED BY
-# THIS LICENSE. THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND
-# CONTRIBUTORS "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
-# LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A
-# PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR
-# CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL,
-# EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO,
-# PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR PROFITS; OR
-# BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER
-# IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE)
-# ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
-# POSSIBILITY OF SUCH DAMAGE.
-
-#
-import tkinter as tk
-from tkinter import ttk
-import logging
-
-# Logger Setup
-p = __name__.split(".")[1]
-logger = logging.getLogger(p)
-
-from aslm.view.custom_widgets.DockableNotebook import DockableNotebook
-
-# Import Sub-Frames
-from aslm.view.main_window_content.camera_display.camera_view.camera_view_tab import (
-    CameraTab,
-)
-from aslm.view.main_window_content.camera_display.camera_view.waveform_tab import (
-    WaveformTab,
-)
-
-
-class CameraNotebook(DockableNotebook):
-    def __init__(self, frame_top_right, *args, **kwargs):
-        # Init notebook
-        DockableNotebook.__init__(self, frame_top_right, *args, **kwargs)
-
-        # Putting notebook 2 into top right frame
-        self.grid(row=0, column=0)
-
-        # Creating the camera tab
-        self.camera_tab = CameraTab(self)
-
-        # Creating the waveform settings tab
-        self.waveform_tab = WaveformTab(self)
-
-        # Tab list
-        tab_list = [self.camera_tab, self.waveform_tab]
-        self.set_tablist(tab_list)
-
-        # Adding tabs to self notebook
-        self.add(self.camera_tab, text="Camera View", sticky=tk.NSEW)
-        self.add(self.waveform_tab, text="Waveform Settings", sticky=tk.NSEW)
