--- src/aslm/view/main_window_content/camera_display/camera_view/waveform_tab.py
@@ -1,100 +1,0 @@
-# Copyright (c) 2021-2022  The University of Texas Southwestern Medical Center.
-# All rights reserved.
-
-# Redistribution and use in source and binary forms, with or without
-# modification, are permitted for academic and research use only (subject to the limitations in the disclaimer below)
-# provided that the following conditions are met:
-
-#      * Redistributions of source code must retain the above copyright notice,
-#      this list of conditions and the following disclaimer.
-
-#      * Redistributions in binary form must reproduce the above copyright
-#      notice, this list of conditions and the following disclaimer in the
-#      documentation and/or other materials provided with the distribution.
-
-#      * Neither the name of the copyright holders nor the names of its
-#      contributors may be used to endorse or promote products derived from this
-#      software without specific prior written permission.
-
-# NO EXPRESS OR IMPLIED LICENSES TO ANY PARTY'S PATENT RIGHTS ARE GRANTED BY
-# THIS LICENSE. THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND
-# CONTRIBUTORS "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
-# LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A
-# PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR
-# CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL,
-# EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO,
-# PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR PROFITS; OR
-# BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER
-# IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE)
-# ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
-# POSSIBILITY OF SUCH DAMAGE.
-
-# Standard Library Imports
-import logging
-
-# Third Party Imports
-import tkinter as tk
-from tkinter import ttk
-from matplotlib.backends.backend_tkagg import FigureCanvasTkAgg
-from matplotlib.figure import Figure
-
-# Local Imports
-from aslm.view.custom_widgets.LabelInputWidgetFactory import LabelInput
-
-# Logger Setup
-p = __name__.split(".")[1]
-logger = logging.getLogger(p)
-
-
-class WaveformTab(tk.Frame):
-    def __init__(self, cam_wave, *args, **kwargs):
-        # Init Frame
-        tk.Frame.__init__(self, cam_wave, *args, **kwargs)
-
-        self.index = 1
-
-        # Formatting
-        tk.Grid.columnconfigure(self, "all", weight=1)
-        tk.Grid.rowconfigure(self, "all", weight=1)
-
-        self.waveform_plots = ttk.Frame(self)
-        self.waveform_plots.grid(row=0, column=0, sticky=tk.NSEW)
-
-        self.fig = Figure(figsize=(6, 6), dpi=100)
-        self.canvas = FigureCanvasTkAgg(self.fig, master=self.waveform_plots)
-        self.canvas.draw()
-
-        self.waveform_settings = WaveformSettings(self)
-        self.waveform_settings.grid(row=1, column=0, sticky=tk.NSEW, padx=5, pady=5)
-
-
-class WaveformSettings(ttk.Labelframe):
-    def __init__(self, wav_view, *args, **kwargs):
-        # Init Frame
-        text_label = "Settings"
-        ttk.Labelframe.__init__(self, wav_view, text=text_label, *args, **kwargs)
-
-        # Formatting
-        tk.Grid.columnconfigure(self, "all", weight=1)
-        tk.Grid.rowconfigure(self, "all", weight=1)
-
-        # Dictionary for widgets
-        self.inputs = {}
-
-        self.inputs["sample_rate"] = LabelInput(
-            parent=self,
-            label="Sample rate",
-            input_class=ttk.Spinbox,
-            input_var=tk.IntVar(),
-            input_args={"from_": 1, "to": 2**16 - 1, "increment": 1, "width": 5},
-        )
-        self.inputs["sample_rate"].grid(row=0, column=0, sticky=tk.NSEW, padx=3, pady=3)
-
-    def get_variables(self):
-        variables = {}
-        for key, widget in self.inputs.items():
-            variables[key] = widget.get()
-        return variables
-
-    def get_widgets(self):
-        return self.inputs
